Refactor eagleye_pp_single_evaluation.py
