+++ eagleye_util/trajectory_plot/scripts/util/calc.py
@@ -23,7 +23,7 @@
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 # SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-# util.py
+# calc.py
 # Author MapIV Hoda
 
 
@@ -750,17 +750,6 @@ def calc_velocity_error(eagleye_velocity,ref_velocity):
     error_velocity['velocity'] = eagleye_velocity - ref_velocity
     return error_velocity
 
-def judge_qual(xyz,qual,qual_num):
-    set_output_df: List[float] = []
-    for i in range(len(xyz)):
-        if qual_num == qual[i]:
-            x = xyz['x'][i]
-            y = xyz['y'][i]
-            z = xyz['z'][i]
-            qual_data = qual[i]
-            set_output_df.append([x,y,z,qual_data])
-    output_df = pd.DataFrame(set_output_df,columns=['x','y','z','qual'])
-    return output_df
 
 def clac_dr(TimeStamp,distance,eagleye_xyz,eagleye_vel_xyz,ref_xyz,distance_length,distance_step):
     last_distance = 0
@@ -795,200 +784,4 @@ def clac_dr(TimeStamp,distance,eagleye_xyz,eagleye_vel_xyz,ref_xyz,distance_leng
                     break
     calc_error = pd.DataFrame(set_calc_error,columns=['start_distance','distance','absolute_pos_x','absolute_pos_y','dr_pos_x','dr_pos_y','error_x','error_y','error_2d'])
     return calc_error
-
-def plot_xyz(ax, x_data, eagleye_enu, rtk_enu, raw_enu, elem, title, y_label):
-    if elem in raw_enu.columns:
-        ax.plot(x_data , raw_enu[elem] , marker=".", linestyle="None",markersize=1, color = "red",  label="gnss raw data")
-    if elem in rtk_enu.columns:
-        ax.plot(x_data , rtk_enu[elem] , marker="o", linestyle="None",markersize=1, color = "green",  label="gnss rtk data")
-    if elem in eagleye_enu.columns:
-        ax.plot(x_data , eagleye_enu[elem] , marker="s", linestyle="None",markersize=1, alpha=0.3, color = "blue",  label="eagleye")
-    ax.set_xlabel('time [s]')
-    ax.set_ylabel(y_label)
-    ax.set_title(title)
-    ax.legend(loc='upper right')
-    ax.grid()
-
-def plot_rpy(ax, x_data, eagleye_plot_df, elem, title, y_label):
-    if elem in eagleye_plot_df.columns:
-        ax.plot(x_data , eagleye_plot_df[elem] , marker="s", linestyle="None",markersize=1, alpha=0.3, color = "blue",  label="eagleye")
-    ax.set_xlabel('time [s]')
-    ax.set_ylabel(y_label)
-    ax.set_title(title)
-    ax.legend(loc='upper right')
-    ax.grid()
-
-
-
-def plot_each(ax, x_data , eagleye, ref, y_data, title, y_label,ref_data_name):
-    if y_data in ref.columns:
-        ax.plot(x_data, ref[y_data] , marker="o", linestyle="None",markersize=1, color = "green",  label=ref_data_name)
-    if y_data in eagleye.columns:
-        ax.plot(x_data , eagleye[y_data] , marker="s", linestyle="None",markersize=1, alpha=0.3, color = "blue",  label="eagleye")
-    ax.set_xlabel('time [s]')
-    ax.set_ylabel(y_label)
-    ax.set_title(title)
-    ax.legend(loc='upper right')
-    ax.grid()
-
-def plot_6DoF_single(x_data, eagleye_enu, rtk_enu, raw_enu, eagleye_plot_df):
-    fig1 = plt.figure()
-    ax_x = fig1.add_subplot(2, 3, 1)
-    ax_y = fig1.add_subplot(2, 3, 2)
-    ax_z = fig1.add_subplot(2, 3, 3)
-    ax_roll = fig1.add_subplot(2, 3, 4)
-    ax_pitch = fig1.add_subplot(2, 3, 5)
-    ax_yaw = fig1.add_subplot(2, 3, 6)
-    plot_xyz(ax_x, x_data, eagleye_enu, rtk_enu, raw_enu, 'x', 'X (East-West)','East [m]')
-    plot_xyz(ax_y, x_data, eagleye_enu, rtk_enu, raw_enu, 'y', 'Y (North-South)','North [m]')
-    plot_xyz(ax_z, x_data, eagleye_enu, rtk_enu, raw_enu, 'z', 'Z (Height)','Height [m]')
-    plot_rpy(ax_roll, x_data, eagleye_plot_df, 'roll', 'Roll' , 'Roll [deg]')
-    plot_rpy(ax_pitch, x_data, eagleye_plot_df, 'pitch', 'Pitch', 'Pitch [deg]')
-    plot_rpy(ax_yaw, x_data, eagleye_plot_df, 'heading', 'Yaw', 'Yaw [deg]')
-
-def plot_6DoF(x_data,eagleye, ref,ref_data_name):
-    fig1 = plt.figure()
-    ax_x = fig1.add_subplot(2, 3, 1)
-    ax_y = fig1.add_subplot(2, 3, 2)
-    ax_z = fig1.add_subplot(2, 3, 3)
-    ax_roll = fig1.add_subplot(2, 3, 4)
-    ax_pitch = fig1.add_subplot(2, 3, 5)
-    ax_yaw = fig1.add_subplot(2, 3, 6)
-    plot_each(ax_x, x_data, eagleye, ref, 'x', 'X (East-West)','East [m]',ref_data_name)
-    plot_each(ax_y, x_data ,eagleye, ref, 'y', 'Y (North-South)','North [m]',ref_data_name)
-    plot_each(ax_z, x_data , eagleye, ref, 'z', 'Z (Height)','Height [m]',ref_data_name)
-    plot_each(ax_roll, x_data , eagleye, ref, 'roll', 'Roll' , 'Roll [deg]',ref_data_name)
-    plot_each(ax_pitch, x_data , eagleye, ref, 'pitch', 'Pitch', 'Pitch [deg]',ref_data_name)
-    plot_each(ax_yaw, x_data , eagleye, ref, 'yaw', 'Yaw', 'Yaw [deg]',ref_data_name)
-
-def plot_each_error(ax, error_data, y_data, title, y_label):
-    ax.plot(error_data['elapsed_time'] , error_data[y_data] , marker="s", linestyle="None",markersize=1, alpha=0.3, color = "blue")
-    ax.set_xlabel('time [s]')
-    ax.set_ylabel(y_label)
-    ax.set_title(title)
-    ax.grid()
-
-def plot_one(ax, error_data, x_data, y_data, title, x_label, y_label, line_style):
-    ax.plot(error_data[x_data] , error_data[y_data] , marker="s", linestyle=line_style, markersize=1, alpha=0.3, color = "blue")
-    ax.set_xlabel(x_label)
-    ax.set_ylabel(y_label)
-    ax.set_title(title)
-    ax.grid()
-
-def plot_two(ax, ref, eagleye, x_data, y_data, title, x_label, y_label, line_style, ref_data_name):
-    ax.plot(ref[x_data] , ref[y_data] , marker=".", linestyle=line_style, markersize=1, color = "red", label=ref_data_name)
-    ax.plot(eagleye[x_data] , eagleye[y_data] , marker="s", linestyle=line_style, markersize=1, alpha=0.3, color = "blue", label="eagleye")
-    ax.set_xlabel(x_label)
-    ax.set_ylabel(y_label)
-    ax.set_title(title)
-    ax.legend(loc='upper right')
-    ax.grid()
-
-def plot_error_6DoF(error_data,ref_data_name):
-    fig2 = plt.figure()
-    fig2.suptitle(ref_data_name + ' - eagleye Error')
-    ax_x = fig2.add_subplot(2, 3, 1)
-    ax_y = fig2.add_subplot(2, 3, 2)
-    ax_z = fig2.add_subplot(2, 3, 3)
-    ax_roll = fig2.add_subplot(2, 3, 4)
-    ax_pitch = fig2.add_subplot(2, 3, 5)
-    ax_yaw = fig2.add_subplot(2, 3, 6)
-    plot_each_error(ax_x, error_data, 'x', 'X (East-West) Error','East error [m]')
-    plot_each_error(ax_y, error_data, 'y', 'Y (North-South) Error','North error [m]')
-    plot_each_error(ax_z, error_data, 'z', 'Z (Height) Error','Height error [m]')
-    plot_each_error(ax_roll, error_data, 'roll', 'Roll Error' , 'Roll error, [deg]')
-    plot_each_error(ax_pitch, error_data, 'pitch', 'Pitch Error', 'Pitch error [deg]')
-    plot_each_error(ax_yaw, error_data, 'yaw', 'Yaw Error', 'Yaw error [deg]')
-
-def plot_error(error_data,ref_data_name):
-    fig4 = plt.figure()
-    fig4.suptitle(ref_data_name + ' - eagleye Error')
-    ax_x = fig4.add_subplot(2, 2, 1)
-    ax_y = fig4.add_subplot(2, 2, 2)
-    ax_across = fig4.add_subplot(2, 2, 3)
-    ax_along = fig4.add_subplot(2, 2, 4)
-    plot_each_error(ax_x, error_data, 'x', 'X (East-West) Error','East error [m]')
-    plot_each_error(ax_y, error_data, 'y', 'Y (North-South) Error','North error [m]')
-    plot_each_error(ax_across, error_data, 'across', 'Across Error','across error [m]')
-    plot_each_error(ax_along, error_data, 'along', 'Along Error','along error [m]')
-
-def plot_error_distributiln(error_data,ref_data_name):
-    fig5 = plt.figure()
-    fig5.suptitle(ref_data_name + ' - eagleye Error')
-    ax_xy = fig5.add_subplot(1, 2, 1)
-    ax_aa = fig5.add_subplot(1, 2, 2)
-    plot_one(ax_xy, error_data, 'x', 'y', 'X-Y error', 'x error [m]', 'y error [m]', 'None')
-    plot_one(ax_aa, error_data, 'across', 'along', 'Across-Along error', 'across error [m]', 'along error [m]', 'None')
-
-def plot_traj(ref_data, data,ref_data_name):
-    fig = plt.figure()
-    ax = fig.add_subplot(1, 1, 1)
-    ax.plot(ref_data['x']-ref_data['x'][0] , ref_data['y']-ref_data['y'][0] , marker=".",linestyle="None",markersize=1, color = "red",  label=ref_data_name)
-    ax.plot(data['x']-ref_data['x'][0] , data['y']-ref_data['y'][0] ,  marker="s",linestyle="None",markersize=1,alpha=0.3 , color = "blue",  label="eagleye")
-    ax.set_xlabel('East [m]')
-    ax.set_ylabel('North [m]')
-    ax.set_title('2D Trajectory')
-    ax.legend(loc='upper right')
-    ax.grid()
-    ax.set_aspect('equal')
-    ax.axis('square')
-
-def plot_traj_three(raw_data, ref_data, data, ref_data_name):
-    fig = plt.figure()
-    ax = fig.add_subplot(1, 1, 1)
-    ax.plot(raw_data['x']-ref_data['x'][0] , raw_data['y']-ref_data['y'][0] , marker=".",linestyle="None",markersize=1, color = "red",  label="gnss raw data")
-    ax.plot(ref_data['x']-ref_data['x'][0] , ref_data['y']-ref_data['y'][0] , marker=".",linestyle="None",markersize=1, color = "green",  label=ref_data_name)
-    ax.plot(data['x']-ref_data['x'][0] , data['y']-ref_data['y'][0] ,  marker="s",linestyle="None",markersize=1,alpha=0.3 , color = "blue",  label="eagleye")
-    ax.set_xlabel('East [m]')
-    ax.set_ylabel('North [m]')
-    ax.set_title('2D Trajectory')
-    ax.legend(loc='upper right')
-    ax.grid()
-    ax.set_aspect('equal')
-    ax.axis('square')
-
-def plot_traj_3d(ref_data, data, ref_data_name):
-    fig = plt.figure()
-    ax = fig.add_subplot(projection='3d')
-    ax.set_title('3D Trajectory')
-    ax.plot3D(ref_data['x'] - ref_data['x'][0] , ref_data['y'] - ref_data['y'][0] ,ref_data['z'] - ref_data['z'][0] , marker=".",linestyle="None",markersize=1, color = "red",label=ref_data_name)
-    ax.plot3D(data['x'] - ref_data['x'][0] , data['y'] - ref_data['y'][0] ,data['z'] - ref_data['z'][0] , marker="s",linestyle="None",markersize=1, alpha=0.3, color = "blue",label="eagleye")
-    ax.set_xlabel('East [m]')
-    ax.set_ylabel('North [m]')
-    ax.set_zlabel('Height [m]')
-    ax.legend(loc='upper right')
-    ax.grid()
-
-def plot_traj_3d_three(raw_data, ref_data, data, ref_data_name):
-    fig = plt.figure()
-    ax = fig.add_subplot(projection='3d')
-    ax.set_title('3D Trajectory')
-    ax.plot3D(raw_data['x'] - ref_data['x'][0] , raw_data['y'] - ref_data['y'][0] ,raw_data['z'] - ref_data['z'][0] , marker=".",linestyle="None",markersize=1, color = "red",label="gnss raw data")
-    ax.plot3D(ref_data['x'] - ref_data['x'][0] , ref_data['y'] - ref_data['y'][0] ,ref_data['z'] - ref_data['z'][0] , marker=".",linestyle="None",markersize=1, color = "green",label=ref_data_name)
-    ax.plot3D(data['x'] - ref_data['x'][0] , data['y'] - ref_data['y'][0] ,data['z'] - ref_data['z'][0] , marker="s",linestyle="None",markersize=1, alpha=0.3, color = "blue",label="eagleye")
-    ax.set_xlabel('East [m]')
-    ax.set_ylabel('North [m]')
-    ax.set_zlabel('Height [m]')
-    ax.legend(loc='upper right')
-    ax.grid()
-
-def plot_traj_qual(xyz, qual):
-    gnss_single = judge_qual(xyz,qual,1)
-    gnss_differential = judge_qual(xyz,qual,2)
-    gnss_fix = judge_qual(xyz,qual,4)
-    gnss_float = judge_qual(xyz,qual,5)
-
-    fig = plt.figure()
-    ax_fix = fig.add_subplot(1, 1, 1)
-    ax_fix.set_title('GNSS positioning solution')
-    ax_fix.plot(gnss_single['x']-xyz['x'][0] , gnss_single['y']-xyz['y'][0] , marker=".",linestyle="None",markersize=1, color = "red",  label="gnss single")
-    ax_fix.plot(gnss_differential['x']-xyz['x'][0] , gnss_differential['y']-xyz['y'][0] , marker=".",linestyle="None",markersize=1, color = "blue",  label="gnss differential")
-    ax_fix.plot(gnss_fix['x']-xyz['x'][0] , gnss_fix['y']-xyz['y'][0] , marker=".",linestyle="None",markersize=1, color = "green",  label="gnss fix")
-    ax_fix.plot(gnss_float['x']-xyz['x'][0] , gnss_float['y']-xyz['y'][0] , marker=".",linestyle="None",markersize=1, color = "yellow",  label="gnss float")    
-    ax_fix.set_xlabel('East [m]')
-    ax_fix.set_ylabel('North [m]')
-    ax_fix.legend(loc='upper right')
-    ax_fix.grid()
-    ax_fix.set_aspect('equal','box')
-    ax_fix.axis('square')
